requirements.txt
dj-database-url
Django
gunicorn
psycopg2
whitenoise
dj-static
Unipath
python-decouple
Pillow
Markdown
bleach
django-material
raven
django-debug-toolbar
django-tables2
django-phonenumber-field